backend/requirements.txt
google-cloud-storage==2.14.0
google-cloud-speech==2.24.0
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
anthropic==0.49.0
python-telegram-bot==21.0
openai==1.60.0
functions-framework==3.5.0
pydub==0.25.1
